Started changing features to improve
