+++ src/main.go
@@ -6,36 +6,38 @@ import (
 	"os"
 )
 
-var allCommands = []commands.Command{
-	commands.Diff,
-	commands.Sync,
-	commands.Init,
-}
-
-var allOptions = []commands.Option{
-	{"help", "h", false},
-	{"force", "f", false},
-	{"timeout", "t", true},
-}
-
 func main() {
+	success := handleCommand()
+
+	if !success {
+		// If no command was given or the command or arguments were invalid, show the general help text.
+		printHelp()
+	}
+}
+
+func handleCommand() bool {
 	positionals, options, hasHelpFlag, err := commands.ParseArgs(os.Args, allOptions)
+
 	if err != nil {
 		// Display the error, print the help text then exit.
-		fmt.Println(err.Error())
-		// If we actually have a sub-command...
-	} else if len(positionals) > 0 {
+		println(err.Error())
+		return false
+	}
+
+	// If we have a sub-command, get it and finds the associated command,
+	// sending associated data to be executed
+	if len(positionals) > 0 {
 		argCmd := positionals[0]
 		for _, cmd := range allCommands {
 			if cmd.Name == argCmd {
 				cmd.Execute(positionals[1:], options, hasHelpFlag) // Pass in all positionals after the sub-command.
-				return                                             // Don't print the help text below.
+				return true                                        // Don't print the help text below.
 			}
 		}
 		fmt.Printf("Invalid command: %s\n", argCmd)
 	}
-	// If no command was given or the command or arguments were invalid, show the general help text.
-	printHelp()
+
+	return false
 }
 
 func printHelp() {
